requirements.txt
# Web server
aiohttp>=3.9.0

# ASR (Speech Recognition)
faster-whisper>=1.0.0      # local Whisper ASR
funasr>=1.0.0              # local FunASR
modelscope                 # FunASR model download

# LLM / Vision client
httpx>=0.27.0

# TTS (Text-to-Speech)
edge-tts>=6.1.0            # online Edge TTS
pyttsx3>=2.90              # local offline TTS (SAPI5 / espeak)

# Audio processing
numpy>=1.24.0
av>=11.0.0                 # PCM resampling for TTS

# OPUS codec
opuslib>=3.0.1

# Config
pyyaml>=6.0

# Image processing (Vision / emoji conversion)
Pillow>=10.0.0

# Alibaba Cloud (optional: AliyunASR, DashScope LLM/Vision)
dashscope>=1.0.0
